refactor(python): share the ES256 wallet JWT builder
The CDP and Openfort wallet tokens duplicated host extraction, request-body
hashing, and the ES256 claim set. Move all three to core/wallet_jwt.py; each
backend keeps only its own key-format parsing.

Openfort now omits reqHash for an empty request body instead of hashing "{}",
matching CDP; the signer always sends a non-empty body.
